refactor(viewport): split EditorViewport into model/ui structure
