+++ package.json
@@ -1,7 +1,7 @@
 {
 	"name": "editor",
 	"private": true,
-	"version": "1.0.0",
+	"version": "1.0.0-beta.2",
 	"type": "module",
 	"scripts": {
 		"dev": "vite",
@@ -37,14 +37,14 @@
 		"@testing-library/jest-dom": "6.9.1",
 		"@testing-library/react": "16.3.0",
 		"@testing-library/user-event": "14.6.1",
-		"@types/node": "^24.7.1",
+		"@types/node": "24.7.1",
 		"@types/react": "19.1.16",
 		"@types/react-dom": "19.1.9",
 		"@types/redux-logger": "3.0.13",
 		"@typescript-eslint/eslint-plugin": "8.45.0",
 		"@typescript-eslint/parser": "8.45.0",
 		"@vitejs/plugin-react-swc": "4.1.0",
-		"@vitest/coverage-istanbul": "^3.2.4",
+		"@vitest/coverage-istanbul": "3.2.4",
 		"@vitest/coverage-v8": "^3.2.4",
 		"@vitest/ui": "3.2.4",
 		"eslint": "9.36.0",
