--- src/entities/editor/model/index.ts
@@ -1,1 +1,0 @@
-export * from '../ui/EditorViewport';
