+++ src/shared/shared/ui/buttons/ViewportButton/index.ts
@@ -1,0 +1,1 @@
+export * from './ViewportButton';
