--- src/widgets/projects-list/index.ts
@@ -1,2 +1,0 @@
-export * from './ProjectList';
-export * from './ProjectSkeleton';
