+++ src/widgets/projects-list/model/index.ts
@@ -1,0 +1,2 @@
+export * from '../ui/ProjectList';
+export * from '../ui/ProjectSkeleton';
