+++ src/widgets/projects-list/ui/ProjectSkeleton.tsx
