+++ src/widgets/projects-list/ui/views/ProjectListEmptyView.tsx
