+++ src/widgets/projects-list/ui/views/__tests_/ProjectListEmptyView.test.tsx
@@ -1,5 +1,5 @@
 import { render, screen } from '@testing-library/react';
-import { ProjectListEmptyView } from '../ProjectListEmptyView';
+import { ProjectListEmptyView } from '../index';
 
 describe('ProjectListEmptyView', () => {
 	it('renders ProjectEmptyState correctly', () => {
