+++ src/widgets/projects-list/ui/views/index.ts
