+++ src/widgets/viewport/model/index.ts
@@ -1,0 +1,2 @@
+export * from '../ui/EditorViewport';
+export * from '../ui/ViewportControls';
